Add more retries before adding an AP to the blacklist
We've encountered environments where due to connectivity issues
wpa_s is not always succesful connecting to the AP. Avoid it being
immediately blacklisted as it used to be and allow 3 attempts
of connection prior to adding to the blacklist.
This change to handle connectivity issues may actually slow down
recovery in load balancing environments.