P2P: enable to set in the conf file the wmm params
In case of P2P GO, the wpa_supplicant uses the
default hostapd parameters. In the default parameters
the ACM bit for video and voice are set to 1, meaning,
P2P devices and stationss which are connected to the
GO cannot pass voice or video data packets.

Signed-hostap: Yoni Divinsky <yoni.divinsky@ti.com>